+++ packages/api/src/index.ts
@@ -9,6 +9,7 @@ export * from "./data/defects";
 export * from "./data/dev";
 export * from "./data/documents";
 export * from "./data/drivers";
+export * from "./data/expoPush";
 export * from "./data/fuelCards";
 export * from "./data/fuelLogs";
 export * from "./data/invites";